tests/baseline/protocol-1.2/inspector/console/console-log-side-effects-expected.txt

CONSOLE MESSAGE: line 17: string
CONSOLE MESSAGE: line 18: 42
CONSOLE MESSAGE: line 19: false
CONSOLE MESSAGE: line 20: undefined
CONSOLE MESSAGE: line 21: null
CONSOLE MESSAGE: line 22: NaN
CONSOLE MESSAGE: line 23: -Infinity
CONSOLE MESSAGE: line 24: 0
CONSOLE MESSAGE: line 25: 42
CONSOLE MESSAGE: line 26: -4.242e-11
CONSOLE MESSAGE: line 27: true
CONSOLE MESSAGE: line 28: foo
CONSOLE MESSAGE: line 29: [object Object]
CONSOLE MESSAGE: line 30: [object Window]
CONSOLE MESSAGE: line 36: [object HTMLParagraphElement]
CONSOLE MESSAGE: line 40: [object Object]
CONSOLE MESSAGE: line 41: 1,2,3
CONSOLE MESSAGE: line 43: [object Uint32Array]
CONSOLE MESSAGE: line 46: 1,2,3,[object Object]
CONSOLE MESSAGE: line 49: [object Object]
CONSOLE MESSAGE: line 52: 1,2,3,[object Object],[object Object]
CONSOLE MESSAGE: line 61: 1,2,3,4,5,,,,,,NaN,,,,,,,NaN,1,2,3,4,5,,
CONSOLE MESSAGE: line 65: [object Object]
CONSOLE MESSAGE: line 68: 1,2,3,[object Object],[object Object],[object Object]
CONSOLE MESSAGE: line 74: 
CONSOLE MESSAGE: line 74: 
CONSOLE MESSAGE: line 74: 
CONSOLE MESSAGE: line 74: 
CONSOLE MESSAGE: line 74: 
CONSOLE MESSAGE: line 77: 
CONSOLE MESSAGE: line 84: 
CONSOLE MESSAGE: line 90: 
Tests various extreme usages of console.log()
